+++ azalea-inventory/src/lib.rs
@@ -1,6 +1,9 @@
 //! Representations of various inventory data structures in Minecraft.
 
+#![feature(min_specialization)]
+
 pub mod components;
+pub mod default_components;
 pub mod item;
 pub mod operations;
 mod slot;