+++ azalea-registry/src/tags/mod.rs
@@ -1,3 +1,4 @@
 pub mod blocks;
+pub mod entities;
 pub mod fluids;
 pub mod items;