+++ azalea/examples/testbot/commands/debug.rs
@@ -10,10 +10,12 @@ use azalea::{
     interact::pick::HitResultComponent,
     packet::game,
     pathfinder::{ExecutingPath, Pathfinder},
+    prelude::ContainerClientExt,
     world::MinecraftEntityId,
 };
 use azalea_core::hit_result::HitResult;
 use azalea_entity::EntityKindComponent;
+use azalea_inventory::components::MaxStackSize;
 use azalea_world::InstanceContainer;
 use bevy_ecs::event::Events;
 use parking_lot::Mutex;
@@ -189,6 +191,16 @@ pub fn register(commands: &mut CommandDispatcher<Mutex<CommandSource>>) {
         let source = ctx.source.lock();
         source.bot.start_use_item();
         source.reply("Ok!");
+        1
+    }));
+    commands.register(literal("maxstacksize").executes(|ctx: &Ctx| {
+        let source = ctx.source.lock();
+        let max_stack_size = source
+            .bot
+            .get_held_item()
+            .get_component::<MaxStackSize>()
+            .map_or(-1, |s| s.count);
+        source.reply(format!("{max_stack_size}"));
         1
     }));