+++ codegen/migrate.py
@@ -1,4 +1,4 @@
-import lib.code.item_components
+import codegen.lib.code.data_components
 import lib.code.inventory
 import lib.code.language
 import lib.code.registry
@@ -15,10 +15,10 @@ import sys
 lib.download.clear_version_cache()
 
 if len(sys.argv) == 1:
-    print('\033[91mYou must provide a version to migrate to.\033[m')
+    print("\033[91mYou must provide a version to migrate to.\033[m")
     version_manifest = lib.download.get_version_manifest()
-    newest_version = version_manifest['latest']['snapshot']
-    print(f'Hint: newest version is \033[1m{newest_version}\033[m')
+    newest_version = version_manifest["latest"]["snapshot"]
+    print(f"Hint: newest version is \033[1m{newest_version}\033[m")
     exit()
 
 
@@ -33,42 +33,49 @@ new_burger_data = lib.extract.get_burger_data_for_version(new_version_id)
 new_packets_report = lib.extract.get_packets_report(new_version_id)
 lib.code.packet.set_packets(new_packets_report)
 
-lib.code.version.set_protocol_version(
-    new_burger_data[0]['version']['protocol'])
+lib.code.version.set_protocol_version(new_burger_data[0]["version"]["protocol"])
 lib.code.version.set_version_name(new_version_id)
 
-print('Updated protocol!')
+print("Updated protocol!")
 
-print('Generating blocks and shapes...')
-new_pumpkin_block_datas = lib.extract.get_pumpkin_data(new_version_id, 'blocks')
+print("Generating blocks and shapes...")
+new_pumpkin_block_datas = lib.extract.get_pumpkin_data(new_version_id, "blocks")
 
 new_block_states_report = lib.extract.get_block_states_report(new_version_id)
 new_registries = lib.extract.get_registries_report(new_version_id)
 new_ordered_blocks = lib.code.blocks.get_ordered_blocks(new_registries)
 
-lib.code.blocks.generate_blocks(new_block_states_report, new_pumpkin_block_datas, new_ordered_blocks, new_burger_data)
+lib.code.blocks.generate_blocks(
+    new_block_states_report,
+    new_pumpkin_block_datas,
+    new_ordered_blocks,
+    new_burger_data,
+)
 lib.code.shapes.generate_block_shapes(new_pumpkin_block_datas, new_block_states_report)
 
-print('Getting en_us.json...')
+print("Getting en_us.json...")
 language = lib.extract.get_en_us_lang(new_version_id)
 lib.code.language.write_language(language)
 
-print('Generating registries...')
+print("Generating registries...")
 import genregistries
+
 genregistries.generate(new_version_id)
 
-print('Generating entity data...')
-burger_entities_data = new_burger_data[0]['entities']
+print("Generating entity data...")
+burger_entities_data = new_burger_data[0]["entities"]
 lib.code.entity.generate_entity_metadata(burger_entities_data, new_mappings)
 lib.code.entity.generate_entity_dimensions(burger_entities_data)
 
-print('Generating item components...')
-lib.code.item_components.generate(new_version_id)
+print("Generating item components...")
+lib.code.data_components.generate(new_version_id)
 
-print('Finishing touches, setting version in README and formatting code...')
+print("Finishing touches, setting version in README and formatting code...")
 lib.code.version.set_version_id(new_version_id)
 
 lib.code.utils.fmt()
 
-print('Done!')
-print('Make sure to `cargo check` and look for the generated `TODO`s to make sure everything is correct!')
+print("Done!")
+print(
+    "Make sure to `cargo check` and look for the generated `TODO`s to make sure everything is correct!"
+)